pyproject.toml
[project]
name = "youtube-copyright-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "google-api-python-client>=2.183.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pyyaml>=6.0.3",
]
